Clarify samples according to parser ideals
Instances of a variable will be automatically converted to:
  (+ A B) -> (+ (get A) (get B))
This will not affect set calls.

For readability, it is much nicer to just use the variable you want
where you want it, instead of having to insert (get N) calls every
time you want to use a variable.
Internally however, these must be converted to (get N) calls.