Start work on the parser, v0
The parser will be implemented in hand-compiled Lithp. This does
make it difficult to update code, but the hope is that a simple
parser can be implemented, and then a more complex parser after
that (v1).

The start of the parser is a library allowing Lithp code to generate
the parsing tree for the Lithp interpreter.

This parser, v0, will work on a fairly strict syntax, and error
messages are likely to not be all that helpful. The idea is to get
a parser implemented in the smallest amount of code, and then work
on a better parser and new language features such as pattern matching.