Move current parser from v0 to v1
It is apparent that the current implementation of the parser is too
complicated. Trying to code a parser in a language under development
is not an easy task.

Instead, the current work is being moved to Platform V1. It will
continue on as Platform V1, whilst a new parser is worked on for
Platform V0.