Bootstrap Parser supports all tested samples.
The Bootstrap Parser has been largely rewritten, and now supports all of
the required constructs for samples such as factorial to work.

See l_src/factorial.lithp for a working example. It can be run with the
new "run.js" script, which will invoke the parser and then run your code.

There may still be cases where the parser does not generate the correct
code, but presently tests pass.
This parser is also fairly lazy, if you miscount your closing braces you
will not be warned of this fact. Your code will likely not run, however.