Can now use custom platform libraries
Platform libraries allow one to extend the interpreter by adding new
runtime builtin functions in JavaScript. This allows projects that use Lithp
to provide necessary runtime functions.

The platform/1 builtin now supports platform libraries.
An example platform library is provided in ext/index.js.

When platform/1 is called, it searches for the given name + /index.js.
This means you can provide a string or an atom to the platform/1 function.